# api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: hello@permify.co
    name: API Support
    url: https://github.com/Permify/permify/issues
  description: Permify is an open source authorization service for creating fine-grained
    and scalable authorization systems.
  license:
    name: Apache-2.0 license
    url: https://github.com/Permify/permify/blob/master/LICENSE
  title: Permify API
  version: v0.9.9
servers:
- url: /
...
paths:
...
  /v1/tenants/{tenant_id}/watch:
    post:
      operationId: watch.watch
      parameters:
      - description: "Identifier of the tenant, if you are not using multi-tenancy\
          \ (have only one tenant) use pre-inserted tenant <code>t1</code> for this\
          \ field. Required, and must match the pattern \\“[a-zA-Z0-9-,]+\\“, max\
          \ 64 bytes."
        in: path
        name: tenant_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/watch_watch_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stream_result_of_WatchResponse'
          description: A successful response.(streaming responses)
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
          description: An unexpected error response.
      summary: watch changes
      tags:
      - Watch
      x-codeSamples:
      - label: go
        lang: go
        source: |
          cr, err := client.Watch.Watch(context.Background(), &v1.WatchRequest{
              TenantId:  "t1",
              SnapToken: "",
          })
          // handle stream response
          for {
              res, err := cr.Recv()

              if err == io.EOF {
                  break
              }

              // res.Changes
          }
      - label: node
        lang: javascript
        source: |
          const permify = require("@permify/permify-node");
          const {WatchResponse} = require("@permify/permify-node/dist/src/grpc/generated/base/v1/service");

          function main() {
              const client = new permify.grpc.newClient({
                  endpoint: "localhost:3478",
              })

              let res = client.watch.watch({
                    tenantId: "t1",
                  snapToken: ""
              })

              handle(res)
          }

          async function handle(res: AsyncIterable<WatchResponse>) {
              for await (const response of res) {
                  // response.changes
              }
          }
      x-codegen-request-body-name: body
      x-content-type: application/json
      x-accepts:
      - application/json
components:
  schemas:
...
    Any:
      additionalProperties:
        type: object
      description: |-
        `Any` contains an arbitrary serialized protocol buffer message along with a
        URL that describes the type of the serialized message.

        Protobuf library provides support to pack/unpack Any values in the form
        of utility functions or additional generated methods of the Any type.

        Example 1: Pack and unpack a message in C++.

            Foo foo = ...;
            Any any;
            any.PackFrom(foo);
            ...
            if (any.UnpackTo(&foo)) {
              ...
            }

        Example 2: Pack and unpack a message in Java.

            Foo foo = ...;
            Any any = Any.pack(foo);
            ...
            if (any.is(Foo.class)) {
              foo = any.unpack(Foo.class);
            }
            // or ...
            if (any.isSameTypeAs(Foo.getDefaultInstance())) {
              foo = any.unpack(Foo.getDefaultInstance());
            }

         Example 3: Pack and unpack a message in Python.

            foo = Foo(...)
            any = Any()
            any.Pack(foo)
            ...
            if any.Is(Foo.DESCRIPTOR):
              any.Unpack(foo)
              ...

         Example 4: Pack and unpack a message in Go

             foo := &pb.Foo{...}
             any, err := anypb.New(foo)
             if err != nil {
               ...
             }
             ...
             foo := &pb.Foo{}
             if err := any.UnmarshalTo(foo); err != nil {
               ...
             }

        The pack methods provided by protobuf library will by default use
        'type.googleapis.com/full.type.name' as the type URL and the unpack
        methods only use the fully qualified type name after the last '/'
        in the type URL, for example "foo.bar.com/x/y.z" will yield type
        name "y.z".

        JSON
        ====
        The JSON representation of an `Any` value uses the regular
        representation of the deserialized, embedded message, with an
        additional field `@type` which contains the type URL. Example:

            package google.profile;
            message Person {
              string first_name = 1;
              string last_name = 2;
            }

            {
              "@type": "type.googleapis.com/google.profile.Person",
              "firstName": <string>,
              "lastName": <string>
            }

        If the embedded message type is well-known and has a custom JSON
        representation, that representation will be embedded adding a field
        `value` which holds the custom JSON in addition to the `@type`
        field. Example (for message [google.protobuf.Duration][]):

            {
              "@type": "type.googleapis.com/google.protobuf.Duration",
              "value": "1.212s"
            }
      example:
        '@type': '@type'
      properties:
        '@type':
          description: |-
            A URL/resource name that uniquely identifies the type of the serialized
            protocol buffer message. This string must contain at least
            one "/" character. The last segment of the URL's path must represent
            the fully qualified name of the type (as in
            `path/google.protobuf.Duration`). The name should be in a canonical form
            (e.g., leading "." is not accepted).

            In practice, teams usually precompile into the binary all types that they
            expect it to use in the context of Any. However, for URLs which use the
            scheme `http`, `https`, or no scheme, one can optionally set up a type
            server that maps type URLs to message definitions as follows:

            * If no scheme is provided, `https` is assumed.
            * An HTTP GET on the URL must yield a [google.protobuf.Type][]
              value in binary format, or produce an error.
            * Applications are allowed to cache lookup results based on the
              URL, or have them precompiled into a binary to avoid any
              lookup. Therefore, binary compatibility needs to be preserved
              on changes to types. (Use versioned type names to manage
              breaking changes.)

            Note: this functionality is not currently available in the official
            protobuf release, and it is not used for type URLs beginning with
            type.googleapis.com. As of May 2023, there are no widely used type server
            implementations and no plans to implement one.

            Schemes other than `http`, `https` (or the empty scheme) might be
            used with implementation specific semantics.
          type: string
      type: object
...
    Attribute:
      description: Attribute represents an attribute of an entity with a specific
        type and value.
      example:
        attribute: attribute
        value:
          '@type': '@type'
        entity:
          id: id
          type: type
      properties:
        entity:
          $ref: '#/components/schemas/Entity'
        attribute:
          title: Name of the attribute
          type: string
        value:
          $ref: '#/components/schemas/Any'
      type: object
...
    DataChange:
      description: "DataChange represents a single change in data, with an operation\
        \ type and the actual change which could be a tuple or an attribute."
      example:
        tuple:
          subject:
            id: id
            type: type
            relation: relation
          entity:
            id: id
            type: type
          relation: relation
        attribute:
          attribute: attribute
          value:
            '@type': '@type'
          entity:
            id: id
            type: type
        operation: OPERATION_UNSPECIFIED
      properties:
        operation:
          $ref: '#/components/schemas/DataChange.Operation'
        tuple:
          $ref: '#/components/schemas/Tuple'
        attribute:
          $ref: '#/components/schemas/Attribute'
      type: object
    DataChange.Operation:
      default: OPERATION_UNSPECIFIED
      description: |2-
         - OPERATION_UNSPECIFIED: Default operation, not specified.
         - OPERATION_CREATE: Creation operation.
         - OPERATION_DELETE: Deletion operation.
      enum:
      - OPERATION_UNSPECIFIED
      - OPERATION_CREATE
      - OPERATION_DELETE
      type: string
    DataChanges:
      description: DataChanges represent changes in data with a snap token and a list
        of data change objects.
      example:
        snap_token: snap_token
        data_changes:
        - tuple:
            subject:
              id: id
              type: type
              relation: relation
            entity:
              id: id
              type: type
            relation: relation
          attribute:
            attribute: attribute
            value:
              '@type': '@type'
            entity:
              id: id
              type: type
          operation: OPERATION_UNSPECIFIED
        - tuple:
            subject:
              id: id
              type: type
              relation: relation
            entity:
              id: id
              type: type
            relation: relation
          attribute:
            attribute: attribute
            value:
              '@type': '@type'
            entity:
              id: id
              type: type
          operation: OPERATION_UNSPECIFIED
      properties:
        snap_token:
          description: The snapshot token.
          type: string
        data_changes:
          description: The list of data changes.
          items:
            $ref: '#/components/schemas/DataChange'
          type: array
      type: object
...
    Entity:
      description: Entity represents an entity with a type and an identifier.
      example:
        id: id
        type: type
      properties:
        type:
          type: string
        id:
          type: string
      type: object
...
    Status:
      example:
        code: 0
        details:
        - '@type': '@type'
        - '@type': '@type'
        message: message
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
        details:
          items:
            $ref: '#/components/schemas/Any'
          type: array
      type: object
    Subject:
      description: "Subject represents an entity subject with a type, an identifier,\
        \ and a relation."
      example:
        id: id
        type: type
        relation: relation
      properties:
        type:
          type: string
        id:
          type: string
        relation:
          type: string
      type: object
...
    Tuple:
      description: "Tuple is a structure that includes an entity, a relation, and\
        \ a subject."
      example:
        subject:
          id: id
          type: type
          relation: relation
        entity:
          id: id
          type: type
        relation: relation
      properties:
        entity:
          $ref: '#/components/schemas/Entity'
        relation:
          type: string
        subject:
          $ref: '#/components/schemas/Subject'
      type: object
...
    WatchResponse:
      description: |-
        WatchResponse is the response message for the Watch RPC. It contains the
        changes in the data that are being watched.
      example:
        changes:
          snap_token: snap_token
          data_changes:
          - tuple:
              subject:
                id: id
                type: type
                relation: relation
              entity:
                id: id
                type: type
              relation: relation
            attribute:
              attribute: attribute
              value:
                '@type': '@type'
              entity:
                id: id
                type: type
            operation: OPERATION_UNSPECIFIED
          - tuple:
              subject:
                id: id
                type: type
                relation: relation
              entity:
                id: id
                type: type
              relation: relation
            attribute:
              attribute: attribute
              value:
                '@type': '@type'
              entity:
                id: id
                type: type
            operation: OPERATION_UNSPECIFIED
      properties:
        changes:
          $ref: '#/components/schemas/DataChanges'
      type: object
...
    watch_watch_request:
      description: |-
        WatchRequest is the request message for the Watch RPC. It contains the
        details needed to establish a watch stream.
      properties:
        snap_token:
          description: "The snap token to avoid stale cache, see more details on [Snap\
            \ Tokens](../../operations/snap-tokens)."
          type: string
      type: object
    Stream_result_of_WatchResponse:
      example:
        result:
          changes:
            snap_token: snap_token
            data_changes:
            - tuple:
                subject:
                  id: id
                  type: type
                  relation: relation
                entity:
                  id: id
                  type: type
                relation: relation
              attribute:
                attribute: attribute
                value:
                  '@type': '@type'
                entity:
                  id: id
                  type: type
              operation: OPERATION_UNSPECIFIED
            - tuple:
                subject:
                  id: id
                  type: type
                  relation: relation
                entity:
                  id: id
                  type: type
                relation: relation
              attribute:
                attribute: attribute
                value:
                  '@type': '@type'
                entity:
                  id: id
                  type: type
              operation: OPERATION_UNSPECIFIED
        error:
          code: 0
          details:
          - '@type': '@type'
          - '@type': '@type'
          message: message
      properties:
        result:
          $ref: '#/components/schemas/WatchResponse'
        error:
          $ref: '#/components/schemas/Status'
      title: Stream result of WatchResponse
      type: object
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: Authorization
      type: apiKey
...